requirements.txt
pywebview==6.2.1
psutil==7.2.2
send2trash==2.1.0
pyinstaller==6.19.0
pillow==11.3.0
